fix: canonicalize trigger event validation and regenerate CLI docs
Address the PR #358 review round:

- Move the trigger event constants and ValidateTriggerEvent into
  model/trigger_event_validate.go so validate.go stays under the 500-line
  file cap.
- Reject surrounding whitespace in ValidateTriggerEvent, passing the raw
  event through from Trigger.Validate. Runtime matching trims both sides,
  so a padded name still matched — the reject keeps stored triggers
  canonical and mirrors the extension ingress contract.
- Close the ext-namespace divergence Greptile found: the extension
  ingress now rejects a bare "ext." event, matching the trigger
  validator, so an accepted activation always has a creatable consumer.
- Regenerate the CLI reference for the --event help text (codegen-check
  CI failure).
