+++ internal/automation/model/target_validate.go
@@ -103,7 +103,7 @@ func (t Trigger) Validate(path string) error {
 	if strings.TrimSpace(t.Event) == "" {
 		return errors.New(nestedPath(path, "event") + " is required")
 	}
-	if err := ValidateTriggerEvent(strings.TrimSpace(t.Event), path); err != nil {
+	if err := ValidateTriggerEvent(t.Event, path); err != nil {
 		return err
 	}
 	if err := ValidateScopeBinding(t.Scope, t.WorkspaceID, path, "workspace_id"); err != nil {
