+++ packages/site/content/docs/cli/automation/triggers/create.mdx
@@ -17,7 +17,7 @@ compozy automation triggers create [flags]
       --agent string                      Agent definition name; choose either --agent or --loop
       --enabled                           Create the trigger enabled or disabled
       --endpoint-slug string              Public endpoint slug for webhook events
-      --event string                      Trigger event name
+      --event string                      Trigger event name: session.created, session.stopped, memory.consolidated, hook.<hook_name>.completed, webhook, or ext.*
       --filter stringArray                Exact-match filter(s): key=value or comma-separated key=value pairs
   -h, --help                              help for create
       --loop string                       Loop name; choose either --loop or --agent
